flask_app/requirements.txt
Flask
requests==2.26.0
openai
python-dotenv
flask_cors
psycopg2-binary